Use floats for delay, improve docs

+++ test/image_test.py
@@ -1369,7 +1369,7 @@ class ImageModuleTest(unittest.TestCase):
     def test_load_animation(self):
         # test loading from a file
         SAMPLE_FRAMES = 10
-        SAMPLE_DELAY = 150
+        SAMPLE_DELAY = 150.0
         SAMPLE_SIZE = (312, 312)
         gif_path = pathlib.Path(example_path("data/animated_sample.gif"))
         for inp in (
@@ -1390,7 +1390,7 @@ class ImageModuleTest(unittest.TestCase):
                     frame, delay = val
                     self.assertIsInstance(frame, pygame.Surface)
                     self.assertEqual(frame.size, SAMPLE_SIZE)
-                    self.assertIsInstance(delay, int)
+                    self.assertIsInstance(delay, float)
                     self.assertEqual(delay, SAMPLE_DELAY)
 
     def test_load_pathlib(self):
